// src/components/spacer/spacer-types.ts
export enum SpacerPreset {
  Shaded = 'shaded',
  BorderedBottom = 'bordered-bottom',
  BorderedLeft = 'bordered-left',
}

export enum SpacerAlign {
  Start = 'start',
  Center = 'center',
  End = 'end',
}

export enum SpacerDirection {
  Row = 'row',
  Column = 'column',
}

export enum SpacerJustify {
  SpaceBetween = 'space-between',
}